Changed: Extract table-post-fixer from TableEditing.
